Start implementing axum server

+++ src/controller/domain/dtos/job.rs
@@ -7,7 +7,7 @@ pub struct Job {
     pub id: Uuid,
     pub name: String,
     pub workflow_id: Uuid,
-    pub threshold: i64,
+    pub threshold: i32,
     pub image: String,
     pub args: Vec<String>,
     pub envs: Vec<String>,
